MIQ-697: Fix Clearing for NoDataReceived trap not received on SNMP sink.
The alert manager sends information on multiple alerts that can be
conceptualized as a JSON object. Example initial portion of such an object:

    {
      "receiver": "webhook-snmp",
      "status": "firing",
      "alerts": [
        {
          "status": "resolved",
          "labels": {
            "adaptor": "EdrFlow",
            "alertname": "NoDataReceived",
            "instance": "192.168.134.246:8090",
            "job": "prometheus-collector",
            "location": "Windsor",
            "monitor": "codelab-monitor",
            "oid": "1.3.6.1.4.1.37140.3.0.1017",
            "service": "collector",
            "severity": "major",
            "status": "process"
            ...

Notice that there is a "status" field at the top-level as well as in the array
containing details for every alert. The web-hook used to reuse the top-level
field for all the alerts, which is sometimes incorrect.